requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
pytesseract
Pillow
